lkl: host ops: fix jump buffer API and implementation
longjmp was called after the function which called setjmp exited (i.e.
lkl_ops->jmp_buf_set), which rendered the context set by setjmp invalid.

To avoid this issue this patch implements a different set jump buffer
API by adding the function to execute as an argument to make sure that
the setjmp context remains valid during the exection of the function.
Thus, calles of the function can safely call longjmp.
