+++ arch/lkl/include/uapi/asm/host_ops.h
@@ -70,6 +70,17 @@ struct lkl_jmp_buf {
  *
  * @gettid - returns the host thread id of the caller, which need not
  * be the same as the handle returned by thread_create
+ *
+ * @jmp_buf_set - runs the give function and setups a jump back point by saving
+ * the context in the jump buffer; jmp_buf_longjmp can be called from the give
+ * function or any callee in that function to return back to the jump back
+ * point
+ *
+ * NOTE: we can't return from jmp_buf_set before calling jmp_buf_longjmp or
+ * otherwise the saved context (stack) is not going to be valid, so we must pass
+ * the function that will eventually call longjmp here
+ *
+ * @jmp_buf_longjmp - perform a jump back to the saved jump buffer
  */
 struct lkl_host_operations {
 	const char *virtio_devices;
@@ -114,7 +125,7 @@ struct lkl_host_operations {
 
 	long (*gettid)(void);
 
-	int (*jmp_buf_set)(struct lkl_jmp_buf *jmpb);
+	void (*jmp_buf_set)(struct lkl_jmp_buf *jmpb, void (*f)(void));
 	void (*jmp_buf_longjmp)(struct lkl_jmp_buf *jmpb, int val);
 };
