+++ tools/lkl/lib/jmp_buf.c
@@ -1,9 +1,10 @@
 #include <setjmp.h>
 #include <lkl_host.h>
 
-int jmp_buf_set(struct lkl_jmp_buf *jmpb)
+void jmp_buf_set(struct lkl_jmp_buf *jmpb, void (*f)(void))
 {
-	return setjmp(*((jmp_buf *)jmpb->buf));
+	if (!setjmp(*((jmp_buf *)jmpb->buf)))
+		f();
 }
 
 void jmp_buf_longjmp(struct lkl_jmp_buf *jmpb, int val)
